Replace ct_server with TesseraCT in setup
In the setup scaffolding workflow, update Fulcio to use the Static CT
log TesseraCT instead of the Trillian-based ct_server.

The createtree job is obsolete, so it is removed.

TesseraCT does not use a config file, all parameters are passed in via
command line, so remove the CT config map and the config field from the
secret.

TesseraCT does not support encrypting the private key, so remove support
for supplying a password.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>